+++ src/Infrastructure/Persistence/Doctrine/Mapping/Condition.Condition.orm.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+  <entity
+    name="App\Domain\Condition\Condition"
+    table="condition">
+    <id name="uuid" type="guid" column="uuid"/>
+    <field name="negate" type="boolean" column="negate" nullable="false"/>
+    <one-to-one field="vehicleCharacteristics" target-entity="App\Domain\Condition\VehicleCharacteristics" mapped-by="regulationCondition"/>
+  </entity>
+</doctrine-mapping>
