--- src/Infrastructure/Persistence/Doctrine/Mapping/Condition.ConditionSet.orm.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity
-    name="App\Domain\Condition\ConditionSet"
-    table="condition_set">
-    <id name="uuid" type="guid" column="uuid"/>
-    <field name="operator" type="string" column="operator" nullable="true" length="5"/>
-    <one-to-one field="regulationCondition" target-entity="App\Domain\Condition\RegulationCondition" inversed-by="conditionSet">
-      <join-column name="regulation_condition_uuid" referenced-column-name="uuid" nullable="false"/>
-    </one-to-one>
-    <one-to-many field="conditions" target-entity="App\Domain\Condition\RegulationCondition" mapped-by="parentConditionSet" />
-  </entity>
-</doctrine-mapping>
